refactor(vc): split vcTreeOps out of visualComponentsSlice
Health-check item 5 — visualComponentsSlice.ts was the largest store slice
(906 LOC, the highest total cognitive complexity of any production file).
The typed domain errors (VisualComponentNameError / ParamNameError /
RecursionError) and the pure page-tree surgery (VC-ref collection, cascade
removal of refs to a deleted VC, subtree cloning for componentization) now
live in a sibling vcTreeOps.ts — the same shape as the earlier
vcSlotReconcile.ts split. The slice keeps only the store actions.

Error-class consumers (ConvertToComponentButton, tests) import from
vcTreeOps; module-size ratchet lowered 924 → 715.
